+++ src/main/java/com/clova/anifriends/domain/recruitment/dto/response/IsAppliedRecruitmentResponse.java
@@ -1,4 +1,4 @@
-package com.clova.anifriends.domain.recruitment.service;
+package com.clova.anifriends.domain.recruitment.dto.response;
 
 public record IsAppliedRecruitmentResponse(boolean isAppliedRecruitment) {
